refactor(runtime): move Crab.md discovery behind Host::discover_instructions
Env::on_before_run used to walk the filesystem directly — reading
~/.crabtalk/Crab.md plus every Crab.md on the path from the
conversation's CWD back to the root. That was the last non-persistence
reach into std::fs from crabtalk-runtime's hot path.

Push the walk behind the Host trait. Host gains a sync
`discover_instructions(&self, cwd: &Path) -> Option<String>` method
with a None default; DaemonHost implements it with the same logic the
runtime used to carry inline. Embedded users who want workspace-aware
instructions override the method on their own Host impl.

Env drops its local `discover_instructions` helper, its `std::fs`
reach, and its `wcore::paths::CONFIG_DIR` dependency for this path.
The only remaining std::fs in runtime/src is the file tools
(os/read, os/edit — intentional, that's what those tools exist for)
and mcp/handler.rs::scan_port_files (the other deferred leak).
